fix bugges

+++ app/Http/Middleware/CheckTimeMiddleware.php
@@ -22,16 +22,18 @@ class CheckTimeMiddleware
         $foods = Food_purchase::where('user_id', $user->id)->where('is_consumed', false)->get();
 
         foreach ($foods as $food) {
-            $expirationTime = Carbon::parse($food->purchased_at)->addHours($food->food->duration_hours);
+            if(!is_null($food->food)){
+                $expirationTime = Carbon::parse($food->purchased_at)->addHours($food->food->duration_hours);
 
-            if (Carbon::now()->greaterThan($expirationTime)) {
-                $dog = UserDog::where('user_id', $user->id)->first();
-                $dog->health_level -= 1;
-                $dog->hunger_level -= 1;
-                $user->balance += $food->food->price * $food->food->income_price;
-                $user->save();
-                $dog->save();
-                $food->delete();
+                if (Carbon::now()->greaterThan($expirationTime)) {
+                    $dog = UserDog::where('user_id', $user->id)->first();
+                    $dog->health_level -= 1;
+                    $dog->hunger_level -= 1;
+                    $user->balance += $food->food->price * $food->food->income_price;
+                    $user->save();
+                    $dog->save();
+                    $food->delete();
+                }
             }
         }
